+++ tests/e2e/suites/basic/basic_suite_test.go
@@ -1,0 +1,76 @@
+package basic
+
+import (
+	"testing"
+	"time"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+
+	"github.com/giantswarm/apptest-framework/pkg/config"
+	"github.com/giantswarm/apptest-framework/pkg/state"
+	"github.com/giantswarm/apptest-framework/pkg/suite"
+
+	"github.com/giantswarm/clustertest/pkg/logger"
+	appsv1 "k8s.io/api/apps/v1"
+	"k8s.io/apimachinery/pkg/types"
+)
+
+const (
+	isUpgrade = false
+)
+
+func TestBasic(t *testing.T) {
+	installNamespace := "default"
+
+	suite.New(config.MustLoad("../../config.yaml")).
+		WithInstallNamespace(installNamespace).
+		WithIsUpgrade(isUpgrade).
+		WithValuesFile("./values.yaml").
+		AfterClusterReady(func() {
+
+			It("should connect to the management cluster", func() {
+				err := state.GetFramework().MC().CheckConnection()
+				Expect(err).NotTo(HaveOccurred())
+			})
+
+			It("should connect to the workload cluster", func() {
+				wcClient, err := state.GetFramework().WC(state.GetCluster().Name)
+				Expect(err).NotTo(HaveOccurred())
+
+				err = wcClient.CheckConnection()
+				Expect(err).NotTo(HaveOccurred())
+			})
+
+		}).
+		BeforeUpgrade(func() {
+
+			It("should not have run the before upgrade", func() {
+				logger.Log("This isn't an upgrade test so this test case shouldn't have happened")
+				Fail("Shouldn't perform pre-upgrade tests if not an upgrade test suite")
+			})
+
+		}).
+		Tests(func() {
+
+			It("has the app running in the cluster", func() {
+				wcClient, err := state.GetFramework().WC(state.GetCluster().Name)
+				Expect(err).NotTo(HaveOccurred())
+
+				Eventually(func() error {
+					logger.Log("Checking if deployment exists in the workload cluster")
+					var dp appsv1.Deployment
+					err := wcClient.Get(state.GetContext(), types.NamespacedName{Namespace: "default", Name: state.GetApplication().AppName}, &dp)
+					if err != nil {
+						logger.Log("Failed to get deployment: %v", err)
+					}
+					return err
+				}).
+					WithPolling(5 * time.Second).
+					WithTimeout(5 * time.Minute).
+					ShouldNot(HaveOccurred())
+			})
+
+		}).
+		Run(t, "Basic Test")
+}